# watch/docs/swagger.yaml
basePath: /
...
info:
  contact: {}
  description: 'Document: ansurfen.github.io/YockNav/'
  title: YockWatch
  version: "1.0"
paths:
...
  /logger/find:
    get:
      consumes:
      - application/json
      description: Get API Document
      parameters:
      - description: File
        in: query
        name: file
        type: string
      - description: Time
        in: query
        name: time
        type: string
      - description: Level
        in: query
        name: level
        type: string
      - description: Caller
        in: query
        name: caller
        type: string
      - description: Message
        in: query
        name: msg
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Get API Document
      tags:
      - Logger
  /logger/parse:
    get:
      consumes:
      - application/json
      description: Get API Document
      parameters:
      - description: Path of logger
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Get API Document
      tags:
      - Logger
...
  /yock/eval:
    post:
      consumes:
      - application/json
      description: Eval Yock Script
      parameters:
      - description: Script
        in: formData
        name: script
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Eval Yock Script
      tags:
      - Yock
...
swagger: "2.0"